+++ src/Dyfrig/Environment.fs
@@ -36,6 +36,8 @@ type Environment =
     val private requestBody      : Stream
     val private responseHeaders  : OwinHeaders
     val private responseBody     : Stream
+    val private callCancelled    : CancellationToken
+    val private owinVersion      : string
 
     val mutable private disposed : bool
 
@@ -48,10 +50,12 @@ type Environment =
             requestBody = unbox dictionary.[Constants.requestBody]
             responseHeaders = unbox dictionary.[Constants.responseHeaders]
             responseBody = unbox dictionary.[Constants.responseBody]
+            callCancelled = unbox dictionary.[Constants.callCancelled]
+            owinVersion = unbox dictionary.[Constants.owinVersion]
         }
 
     /// Initializes a new Environment from parameters, adding defaults for optional response parameters.
-    new (requestMethod: string, requestScheme: string, requestPathBase: string, requestPath: string, requestQueryString: string, requestProtocol: string, requestHeaders, ?requestBody, ?responseHeaders, ?responseBody) as x =
+    new (requestMethod: string, requestScheme: string, requestPathBase: string, requestPath: string, requestQueryString: string, requestProtocol: string, requestHeaders, ?requestBody, ?responseHeaders, ?responseBody, ?callCancelled) as x =
         // TODO: Consider parsing the URI rather than requiring the pieces to be passed in explicitly.
         {
             inherit Dictionary<string, obj>(StringComparer.Ordinal)
@@ -60,6 +64,8 @@ type Environment =
             requestBody = defaultArg requestBody Stream.Null
             responseHeaders = defaultArg responseHeaders (new Dictionary<_,_>(HashIdentity.Structural) :> IDictionary<_,_>)
             responseBody = defaultArg responseBody (new MemoryStream() :> Stream)
+            callCancelled = defaultArg callCancelled (let cts = new CancellationTokenSource() in cts.Token)
+            owinVersion = "1.0"
         }
         then do
             x.Add(Constants.requestMethod, requestMethod)
@@ -72,6 +78,8 @@ type Environment =
             x.Add(Constants.requestBody, x.requestBody)
             x.Add(Constants.responseHeaders, x.responseHeaders)
             x.Add(Constants.responseBody, x.responseBody)
+            x.Add(Constants.callCancelled, x.callCancelled)
+            x.Add(Constants.owinVersion, x.owinVersion)
 
     /// Gets a value with the specified key from the environment dictionary as the specified type 'a.
     static member inline Get<'a> (environment: OwinEnv, key: string) =
@@ -170,6 +178,14 @@ type Environment =
     /// Gets the response body stream.
     member x.ResponseBody = x.responseBody
 
+    /// Gets or sets the `CancellationToken` indicating whether the request has been cancelled.
+    member x.CallCancelled
+        with get() : int = unbox x.[Constants.callCancelled]
+        and set(v : int) = x.[Constants.callCancelled] <- v
+
+    /// Gets the current OWIN version.
+    member x.OwinVersion = x.owinVersion
+
     /// Overridable disposal implementation for this instance.
     abstract Dispose : bool -> unit
     default x.Dispose(disposing) =
@@ -192,6 +208,7 @@ type Environment =
 /// Helper functions for working with an OWIN environment dictionary
 [<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
 module Environment =
+
     /// Returns an Environment from an `OwinEnv`.
     /// If the dictionary is already an Environment, then the instance is cast and returned
     /// rather than mapped into a new instance.
@@ -200,3 +217,71 @@ module Environment =
         match environment with
         | :? Environment as e -> e
         | _ as d -> new Environment(d)
+    
+/// Helper functions for working with an OWIN environment dictionary
+[<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
+module SystemNetHttpAdapter =
+
+    open System.Diagnostics.Contracts
+    open System.Net.Http
+    open Environment
+
+    /// Literal defintion for the `dyfrig.Environment` key.
+    [<Literal>]
+    let dyfrigEnvironment = "dyfrig.Environment"
+
+    /// Converts the `OwinEnv` into an `HttpRequestMessage`.
+    [<CompiledName("ToHttpRequestMesage")>]
+    let toHttpRequestMessage (environment: OwinEnv) =
+        let env = environment |> toEnvironment
+        match env.GetRequestUri() with
+        | Some requestUri ->
+            let content = new StreamContent(env.RequestBody)
+            let request = new HttpRequestMessage(HttpMethod(env.RequestMethod), requestUri, Content = content)
+            for header in env.RequestHeaders.Keys do
+                request.Headers.TryAddWithoutValidation(header, env.RequestHeaders.[header]) |> ignore
+            request.Version <- Version(env.RequestProtocol.Substring(5))
+            request.Properties.Add(dyfrigEnvironment, env)
+            request.Properties.Add(Constants.callCancelled, env.CallCancelled)
+            // TODO: Add additional, common properties here.
+            Some request
+        | None -> None
+    
+    /// Invokes an `HttpResponseMessage` in an OWIN handler.
+    [<CompiledName("InvokeHttpResponseMessage")>]
+    let invokeHttpResponseMessage (response: HttpResponseMessage) =
+        Contract.Requires(response.RequestMessage <> null)
+        Contract.Requires(response.RequestMessage.Properties.ContainsKey(dyfrigEnvironment))
+        Contract.Requires(response.RequestMessage.Properties.[dyfrigEnvironment] <> null)
+        Contract.Requires((response.RequestMessage.Properties.[dyfrigEnvironment] :?> Environment).ResponseBody <> null)
+
+        let env = response.RequestMessage.Properties.[dyfrigEnvironment] :?> Environment
+        env.ResponseStatusCode <- int response.StatusCode
+        env.ResponseReasonPhrase <- response.StatusCode.ToString()
+        for header in response.Headers do
+            env.ResponseHeaders.Add(header.Key, header.Value |> Seq.toArray)
+        response.Content.CopyToAsync(env.ResponseBody)
+
+    /// Adapts a function of type `HttpRequestMessage -> Async<HttpResponseMessage>` to an OWIN handler.
+    [<CompiledName("FromAsyncSystemNetHttp")>]
+    let fromAsyncSystemNetHttp (f: HttpRequestMessage -> Async<HttpResponseMessage>) =
+        OwinAppFunc(fun env ->
+            let request = env |> toHttpRequestMessage |> Option.get
+            async {
+                let! response = f request
+                do! (invokeHttpResponseMessage response).ContinueWith(Func<_,_>(fun _ -> ())) |> Async.AwaitTask
+            }
+            |> Async.StartAsTask
+            :> Task)
+
+    /// Adapts a function of type `HttpRequestMessage -> Task<HttpResponseMessage>` to an OWIN handler.
+    [<CompiledName("FromSystemNetHttp")>]
+    let fromSystemNetHttp (f: HttpRequestMessage -> Task<HttpResponseMessage>) =
+        OwinAppFunc(fun env ->
+            let request = env |> toHttpRequestMessage |> Option.get
+            async {
+                let! response = f request |> Async.AwaitTask
+                do! (invokeHttpResponseMessage response).ContinueWith(Func<_,_>(fun _ -> ())) |> Async.AwaitTask
+            }
+            |> Async.StartAsTask
+            :> Task)