+++ RELEASE_NOTES.md
@@ -1,4 +1,8 @@
-### New in 0.2.0 - (TBD)
+### New in 0.3.0 - (Released 2014/05/27)
+* Added mapping to System.Net.Http types.
+
+### New in 0.2.0 - (Released 2014/05/26)
+* Added additional members through which to retrieve the request URI and the base URI.
 
 ### New in 0.1.131128 - (Released 2013/11/28)
 * First official release.